update version, update readme, add repo to package json, add required exports, rename control model files
