+++ libs/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ngx-fast-forms/core",
-  "version": "0.0.1-rc1",
+  "version": "0.0.1-rc2",
   "peerDependencies": {
     "@angular/common": "^14.0.0",
     "@angular/core": "^14.0.0",
@@ -8,5 +8,9 @@
   },
   "dependencies": {
     "tslib": "^2.3.0"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Micky002/angular-fast-forms"
   }
 }
