+++ libs/material/package.json
@@ -1,11 +1,17 @@
 {
   "name": "@ngx-fast-forms/material",
-  "version": "0.0.1-rc1",
+  "version": "0.0.1-rc2",
   "peerDependencies": {
     "@angular/common": "^14.0.0",
-    "@angular/core": "^14.0.0"
+    "@angular/core": "^14.0.0",
+    "@angular/forms": "^14.0.0",
+    "@angular/material": "^14.0.0"
   },
   "dependencies": {
     "tslib": "^2.3.0"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Micky002/angular-fast-forms"
   }
 }
