+++ libs/material/src/index.ts
@@ -1,1 +1,5 @@
+export * from './lib/ui/input/input.models';
+
+export * from './lib/ui/select/select.models';
+
 export * from './lib/material-fast-forms.module';
