--- libs/material/src/lib/ui/select/select-option.model.ts
@@ -1,4 +1,0 @@
-export interface SelectOption {
-  value: string | number;
-  name: string;
-}
