+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "angular-fast-forms",
-  "version": "0.0.1-rc1",
+  "version": "0.0.1-rc2",
   "license": "MIT",
   "scripts": {
     "ng": "nx",
@@ -47,6 +47,7 @@
     "@angular/cli": "14.1.3",
     "@angular/compiler-cli": "14.1.3",
     "@angular/language-service": "14.1.3",
+    "@jscutlery/semver": "^2.26.0",
     "@nrwl/cli": "14.5.9",
     "@nrwl/cypress": "14.5.9",
     "@nrwl/eslint-plugin-nx": "14.5.9",
